requirements.txt
nodriver>=0.38
python-dotenv>=1.0
sib-api-v3-sdk>=7.6
flask>=3.0
requests>=2.31
Pillow>=10.0
